requirements.txt
# FastAPI i servidor
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# OCR
pytesseract>=0.3.13
google-cloud-vision>=3.8.0
Pillow>=11.0.0
opencv-python-headless>=4.10.0
numpy>=2.0.0
scipy>=1.14.0

# Utils
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
